+++ fhir/lib/primitive_types/base64binary.dart
@@ -6,7 +6,7 @@ import 'dart:convert';
 // Package imports:
 import 'package:yaml/yaml.dart';
 
-//
+import 'primitive_type_exceptions.dart';
 
 class Base64Binary {
   const Base64Binary._(
@@ -33,7 +33,7 @@ class Base64Binary {
       ? Base64Binary.fromJson(jsonDecode(jsonEncode(loadYaml(yaml))))
       : yaml is YamlMap
           ? Base64Binary.fromJson(jsonDecode(jsonEncode(yaml)))
-          : throw FormatException(
+          : throw YamlFormatException<Base64Binary>(
               'FormatException: "$json" is not a valid Yaml string or YamlMap.');
 
   final String _valueString;