+++ fhir/lib/primitive_types/boolean.dart
@@ -6,6 +6,8 @@ import 'dart:convert';
 // Package imports:
 import 'package:yaml/yaml.dart';
 
+import 'primitive_type_exceptions.dart';
+
 //
 
 class Boolean {
@@ -20,8 +22,9 @@ class Boolean {
           ? Boolean._(inValue, inValue.toLowerCase() == 'true', true, false)
           : Boolean._(inValue, null, false, false);
     } else {
-      throw ArgumentError(
-          "Boolean cannot be constructed from $inValue of type '${inValue.runtimeType.toString()}'.");
+      throw CannotBeConstructed<Boolean>(
+          'Boolean cannot be constructed from $inValue of type '
+          "'${inValue.runtimeType.toString()}'.");
     }
   }
 
@@ -31,7 +34,7 @@ class Boolean {
       ? Boolean.fromJson(jsonDecode(jsonEncode(loadYaml(yaml))))
       : yaml is YamlMap
           ? Boolean.fromJson(jsonDecode(jsonEncode(yaml)))
-          : throw FormatException(
+          : throw YamlFormatException<Boolean>(
               'FormatException: "$json" is not a valid Yaml string or YamlMap.');
 
   final String _valueString;