+++ fhir/lib/primitive_types/fhir_url.dart
@@ -5,6 +5,7 @@ import 'dart:convert';
 
 // Package imports:
 import 'package:yaml/yaml.dart';
+import 'primitive_type_exceptions.dart';
 
 class FhirUrl {
   const FhirUrl._(this._valueString, this._valueUri, this._isValid);
@@ -16,7 +17,8 @@ class FhirUrl {
       final Uri? tempUri = Uri.tryParse(inValue);
       return FhirUrl._(inValue, tempUri, tempUri != null);
     }
-    throw ArgumentError('FhirUrl cannot be constructed from $inValue.');
+    throw CannotBeConstructed<FhirUrl>(
+        'FhirUrl cannot be constructed from $inValue.');
   }
 
   factory FhirUrl.fromJson(dynamic json) => FhirUrl(json);
@@ -25,7 +27,7 @@ class FhirUrl {
       ? FhirUrl.fromJson(jsonDecode(jsonEncode(loadYaml(yaml))))
       : yaml is YamlMap
           ? FhirUrl.fromJson(jsonDecode(jsonEncode(yaml)))
-          : throw FormatException(
+          : throw YamlFormatException<FhirUrl>(
               'FormatException: "$json" is not a valid Yaml string or YamlMap.');
 
   final String _valueString;