+++ fhir/lib/primitive_types/id.dart
@@ -6,6 +6,7 @@ import 'dart:convert';
 // Package imports:
 import 'package:uuid/uuid.dart' as uuid;
 import 'package:yaml/yaml.dart';
+import 'primitive_type_exceptions.dart';
 
 /// returns the same resource with a new ID (even if there is already an ID present)
 Id newId() => Id(const uuid.Uuid().v4());
@@ -24,7 +25,7 @@ class Id {
       ? Id.fromJson(jsonDecode(jsonEncode(loadYaml(yaml))))
       : yaml is YamlMap
           ? Id.fromJson(jsonDecode(jsonEncode(yaml)))
-          : throw FormatException(
+          : throw YamlFormatException<Id>(
               'FormatException: "$json" is not a valid Yaml string or YamlMap.');
 
   final String _valueString;