+++ fhir/lib/primitive_types/integer.dart
@@ -6,6 +6,7 @@ import 'package:yaml/yaml.dart';
 
 // Project imports:
 import 'fhir_number.dart';
+import 'primitive_type_exceptions.dart';
 
 class Integer extends FhirNumber {
   const Integer._(String valueString, int? valueNumber, bool isValid)
@@ -18,7 +19,8 @@ class Integer extends FhirNumber {
       return Integer._(inValue.toString(), int.tryParse(inValue.toString()),
           int.tryParse(inValue.toString()) != null);
     }
-    throw ArgumentError('Integer cannot be constructed from $inValue.');
+    throw CannotBeConstructed<Integer>(
+        'Integer cannot be constructed from $inValue.');
   }
 
   factory Integer.fromJson(dynamic json) => Integer(json);
@@ -27,7 +29,7 @@ class Integer extends FhirNumber {
       ? Integer.fromJson(jsonDecode(jsonEncode(loadYaml(yaml))))
       : yaml is YamlMap
           ? Integer.fromJson(jsonDecode(jsonEncode(yaml)))
-          : throw FormatException(
+          : throw YamlFormatException<Integer>(
               'FormatException: "$json" is not a valid Yaml string or YamlMap.');
 
   int? get value => valueNumber as int?;