+++ fhir/lib/primitive_types/uuid.dart
@@ -5,6 +5,7 @@ import 'dart:convert';
 
 // Package imports:
 import 'package:yaml/yaml.dart';
+import 'primitive_type_exceptions.dart';
 
 class Uuid {
   const Uuid._(this._valueString, this._valueUri, this._isValid);
@@ -22,7 +23,7 @@ class Uuid {
       ? Uuid.fromJson(jsonDecode(jsonEncode(loadYaml(yaml))))
       : yaml is YamlMap
           ? Uuid.fromJson(jsonDecode(jsonEncode(yaml)))
-          : throw FormatException(
+          : throw YamlFormatException<Uuid>(
               'FormatException: "$json" is not a valid Yaml string or YamlMap.');
 
   final String _valueString;